Removed unused InventoryAPI from KafkaEnabledInventoryService
The Kafka based inventory service did not need the API class instance
so common code was moved into an abstract base class and the API is only
used in the DefaultInventoryServiceImplementation
