+++ src/main/java/org/candlepin/insights/inventory/DefaultInventoryService.java
@@ -20,23 +20,16 @@
  */
 package org.candlepin.insights.inventory;
 
-import org.candlepin.insights.api.model.ConsumerInventory;
-import org.candlepin.insights.api.model.OrgInventory;
 import org.candlepin.insights.exception.inventory.InventoryServiceException;
 import org.candlepin.insights.inventory.client.model.BulkHostOut;
 import org.candlepin.insights.inventory.client.model.CreateHostIn;
-import org.candlepin.insights.inventory.client.model.FactSet;
 import org.candlepin.insights.inventory.client.resources.HostsApi;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.time.OffsetDateTime;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 
@@ -63,7 +56,7 @@ import java.util.stream.Collectors;
  * The visit methods can then get called at the appropriate places in sendHostUpdate and createHost
  * allowing us to externalize manipulations to the implementation(s) of ConduitVisitor.
  */
-public class DefaultInventoryService implements InventoryService {
+public class DefaultInventoryService extends InventoryService {
     private static final Logger log = LoggerFactory.getLogger(DefaultInventoryService.class);
 
     private final HostsApi hostsInventoryApi;
@@ -90,57 +83,4 @@ public class DefaultInventoryService implements InventoryService {
         }
     }
 
-    /**
-     * Given a set of facts, report them as a host to the inventory service.
-     *
-     * @return the new host.
-     */
-    protected CreateHostIn createHost(ConduitFacts conduitFacts, OffsetDateTime syncTimestamp) {
-        Map<String, Object> rhsmFactMap = new HashMap<>();
-        rhsmFactMap.put("orgId", conduitFacts.getOrgId());
-        if (conduitFacts.getCpuSockets() != null) {
-            rhsmFactMap.put("CPU_SOCKETS", conduitFacts.getCpuSockets());
-        }
-        if (conduitFacts.getCpuCores() != null) {
-            rhsmFactMap.put("CPU_CORES", conduitFacts.getCpuCores());
-        }
-        if (conduitFacts.getMemory() != null) {
-            rhsmFactMap.put("MEMORY", conduitFacts.getMemory());
-        }
-        if (conduitFacts.getArchitecture() != null) {
-            rhsmFactMap.put("ARCHITECTURE", conduitFacts.getArchitecture());
-        }
-        if (conduitFacts.getIsVirtual() != null) {
-            rhsmFactMap.put("IS_VIRTUAL", conduitFacts.getIsVirtual());
-        }
-        if (conduitFacts.getVmHost() != null) {
-            rhsmFactMap.put("VM_HOST", conduitFacts.getVmHost());
-        }
-        if (conduitFacts.getRhProd() != null) {
-            rhsmFactMap.put("RH_PROD", conduitFacts.getRhProd());
-        }
-
-        rhsmFactMap.put("SYNC_TIMESTAMP", syncTimestamp);
-
-        FactSet rhsmFacts = new FactSet()
-            .namespace("rhsm")
-            .facts(rhsmFactMap);
-        List<FactSet> facts = new LinkedList<>();
-        facts.add(rhsmFacts);
-
-        CreateHostIn host = new CreateHostIn();
-        host.setAccount(conduitFacts.getAccountNumber());
-        host.setFqdn(conduitFacts.getFqdn());
-        host.setSubscriptionManagerId(conduitFacts.getSubscriptionManagerId());
-        host.setBiosUuid(conduitFacts.getBiosUuid());
-        host.setIpAddresses(conduitFacts.getIpAddresses());
-        host.setMacAddresses(conduitFacts.getMacAddresses());
-        host.facts(facts);
-        return host;
-    }
-
-    public OrgInventory getInventoryForOrgConsumers(List<ConduitFacts> conduitFactsForOrg) {
-        List<ConsumerInventory> hosts = new ArrayList<>(conduitFactsForOrg);
-        return new OrgInventory().consumerInventories(hosts);
-    }
 }
